deduplicate cubic_farthest_fit_inside and split_cubic_at across transform modules
Move shared TOLERANCE, Cubic type alias, split_cubic_at, and
cubic_farthest_fit_inside into transform/mod.rs as pub(super) items.
Also flatten the LineTo merge condition in merge_curves.rs to eliminate
a redundant result_starts re-fetch.
